+++ packages/harmony/src/assets/styles/border-radius.css
@@ -8,5 +8,5 @@ root {
   --harmony-border-radius-m: 8px;
   --harmony-border-radius-l: 12px;
   --harmony-border-radius-xl: 16px;
-  --harmony-border-radius-2xl: 320px;
+  --harmony-border-radius-2xl: 32px;
 }
