+++ packages/web/src/pages/sign-in-page/SignInPageMobile.module.css
@@ -1,0 +1,22 @@
+.root {
+  background: radial-gradient(
+      77.16% 77.16% at 50% 51.81%,
+      rgba(91, 35, 225, 0.8) 0%,
+      rgba(113, 41, 230, 0.64) 67.96%,
+      rgba(162, 47, 235, 0.5) 100%
+    ),
+    url('../../assets/img/2-DJ-4-3.jpg'), lightgray 50% / cover no-repeat;
+  background-position-y: bottom;
+  background-position-x: -110px;
+  background-size: 150%;
+}
+
+.content {
+  background-color: var(--harmony-bg-white);
+  border-bottom-left-radius: var(--harmony-border-radius-2xl);
+  border-bottom-right-radius: var(--harmony-border-radius-2xl);
+}
+
+.createAccountRow {
+  color: var(--harmony-white);
+}
