+++ packages/web/src/pages/sign-in-page/SignInPageMobile.tsx
@@ -1,0 +1,93 @@
+import {
+  Box,
+  Button,
+  Flex,
+  IconArrowRight,
+  IconAudiusLogoHorizontalColor,
+  IconVisibilityHidden,
+  Text
+} from '@audius/harmony'
+import { Form } from 'formik'
+import { Link } from 'react-router-dom'
+
+import { HarmonyTextField } from 'components/form-fields/HarmonyTextField'
+
+import styles from './SignInPageMobile.module.css'
+
+const messages = {
+  title: 'Sign Into Audius',
+  emailLabel: 'Email',
+  passwordLabel: 'Password',
+  signIn: 'Sign In',
+  newToAudius: 'New to Audius?',
+  createAccount: 'Create an Account',
+  forgotPassword: 'Forgot password?'
+}
+
+export const SignInPageMobile = () => {
+  return (
+    <Flex
+      className={styles.root}
+      direction='column'
+      w='100%'
+      h='100%'
+      justifyContent='space-between'
+      pb='4xl'
+    >
+      <Flex
+        className={styles.content}
+        w='100%'
+        pv='2xl'
+        ph='l'
+        direction='column'
+        gap='2xl'
+        justifyContent='space-between'
+      >
+        <Flex direction='column' gap='2xl' alignItems='center'>
+          <IconAudiusLogoHorizontalColor />
+          <Text variant='heading' size='l' tag='h1' color='heading'>
+            {messages.title}
+          </Text>
+          <Box w='100%'>
+            <Form>
+              <Flex direction='column' gap='2xl' w='100%'>
+                <Flex direction='column' gap='l'>
+                  {/* TODO: replace old TextField */}
+                  <HarmonyTextField name='email' label={messages.emailLabel} />
+                  {/* TODO: password visibility icon and toggle */}
+                  <HarmonyTextField
+                    name='password'
+                    label={messages.passwordLabel}
+                    endIcon={IconVisibilityHidden}
+                    type='password'
+                  />
+                </Flex>
+                <Flex direction='column' gap='l'>
+                  <Button iconRight={IconArrowRight} type='submit'>
+                    {messages.signIn}
+                  </Button>
+                  <Text color='heading' variant='body'>
+                    {/* TODO: link destination */}
+                    {messages.forgotPassword}
+                  </Text>
+                </Flex>
+              </Flex>
+            </Form>
+          </Box>
+        </Flex>
+      </Flex>
+      <Flex
+        className={styles.createAccountRow}
+        direction='row'
+        w='100%'
+        justifyContent='center'
+        gap='xs'
+        mb='4xl'
+      >
+        {/* TODO: args look good but style doesn't match design */}
+        <Text variant='title'>{messages.newToAudius}</Text>
+        <Link to={''}>{messages.createAccount}</Link>
+      </Flex>
+    </Flex>
+  )
+}
